[Refactor] #8 코드 리팩토링
